+++ web/app/themes/lark-child/config/theme.php
@@ -1,0 +1,22 @@
+<?php
+
+return [
+
+    /*
+    |--------------------------------------------------------------------------
+    | Fonts
+    |--------------------------------------------------------------------------
+    |
+    | Predefined font classes, these can all be override if need be.
+    |
+    */
+
+    'fonts' => [
+        'h1' => [
+            'headingSize' => 'h1',
+            'font' => 'font-cardo font-weight-bold',
+            'margin' => 'mb-30',
+        ],
+    ],
+
+];